% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_pilot.R
\name{auto_pilot}
\alias{auto_pilot}
\title{Auto-pilot prediction of allosteric hub residues from a MD simulation.}
\usage{
auto_pilot(workingdir, traj.length)
}
\arguments{
\item{workingdir}{Path to the input file (eg. "~/sirpaulnurse/MD_dir/")}

\item{traj.length}{Length of the molecular dynamics simulation (in ns)}
}
\value{
Predicted allosteric hub residues
}
\description{
This function reads the output of a molecular dynamics engine (in *.xtc format)
and a topology file (in *.pdb format), and runs through all the analysis functions
of AlloHubMat automatically. Where user-defined parameters are required,
default values are used.
Output files are stored in an automatically-generated directory.
}
