% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_eigensystem.R
\name{comp_eigensystem}
\alias{comp_eigensystem}
\title{Compute the eigensystem for the list of mutual information matrices}
\usage{
comp_eigensystem(MImatlist)
}
\arguments{
\item{MImatlist}{A list of mutual information matrices}
}
\value{
A list containing the eigensystems for each of the mutual information matrices
in the input list
}
\description{
This function computes the covariance matrix and eigensystems for each of
the mutual information matrices, provided to the function, in the input list.
}
