% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_overlap.R
\name{covar_overlap}
\alias{covar_overlap}
\title{Compute the covariance overlap between two matrices}
\usage{
covar_overlap(esA, irange, esB, jrange)
}
\arguments{
\item{esA}{First eigen-system}

\item{irange}{Eigenvalue range of first eigen-system}

\item{esB}{Second eigen-system}

\item{jrange}{Eigenvalue range of second eigen-system}
}
\value{
Return the covariance overlap between the two input eigen-systems
	   for the defined ranges of eigenvalues
}
\description{
This function computes the covariance overlap between two matrices of identical
dimensions.
}
\details{
The matrix calculation is performed according to the following equation (LaTex):
		\begin{equation}
			\Omega_{A:B} = 1 - \lbrace \frac{\sum_{i=1}^{3N}(\lambda_{i}^{A} + \lambda_{i}^{B}) - 2 \sum_{i=1}^{3N} \sum_{j=1}^{3N} (\lambda_{i}^{A} \lambda_{i}^{B})^{0.5} (\textbf{v}_{i}^{A} \cdot \textbf{v}_{i}^{B})^{2}}{\sum_{i=1}^{3N}(\lambda_{i}^{A} + \lambda_{i}^{B})} \rbrace ^{0.5}
		\end{equation}
}
