% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sectors.R
\name{detect_sectors}
\alias{detect_sectors}
\title{Automated detection of ergodic sectors from a list of time-averaged mutual information matrices.}
\usage{
detect_sectors(overlap.matrix.s, sBlock)
}
\arguments{
\item{overlap.matrix.s}{Smoothed matrix of covariance overlaps, computed using matrix_smooth()}

\item{sBlock}{trajectory block size (in ns)}
}
\value{
Returns a dataframe containing a summary of any detected ergodic sectors, including the 
time in the MD trajectory in which they occur
}
\description{
This function takes an input matrix and extract so-called ergodic sectors. The input is a single
matrix generated by matrix_smooth(), and returns a summary of any detected ergodic sectors, and
when they occur in the MD simulation.
}
