% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_smooth.R
\name{matrix_smooth}
\alias{matrix_smooth}
\title{Smooth the matrix of covariance overlaps using a kernal smoothing algorithm.}
\usage{
matrix_smooth(overlap.matrix)
}
\arguments{
\item{overlap.matrix}{Matrix of covariance overlaps, computed using block_overlap()}
}
\value{
A matrix containing the covariance overlap between each of the trajectory blocks
}
\description{
This function takes an input matrix and performs a kernal smoothing.
The input is a single
matrix and returns a single matrix with the same dimensions as that of the input matrix.
}
