% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{read_traj_file}
\alias{read_traj_file}
\title{Read input trajectory}
\usage{
read_traj_file(x, y, a, b, ...)
}
\arguments{
\item{Input}{file name (including path) and format ("dcd" or "xtc").}
}
\value{
.
}
\description{
\code{read_traj_file}
  reads a molecular trajectory file in DCD or XTC format.
}
\examples{
  trajectory_object = read_traj_file(traj_filename, traj_format, start, end

}
