% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sa_encode}
\alias{sa_encode}
\title{Encode input trajectory}
\usage{
sa_encode(x, y, z, ...)
}
\arguments{
\item{Input}{file name (including path) and format ("dcd" or "xtc").}
}
\value{
SA trajectory.
}
\description{
\code{sa_encode}
  encodes a trajectory (DCD or XTC format) into a structural string.
}
\examples{
  sa_trajectory = sa_encode(sadata, trajectory, format)

}
