% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sectors.R
\name{sectors_2dplot}
\alias{sectors_2dplot}
\title{2-dimensional plot of the covariance overlap between continguous mutual information matrices, defined in the input trajectory.}
\usage{
sectors_2dplot(overlap.matrix.s, t.traj)
}
\arguments{
\item{overlap.matrix.s}{Smoothed matrix of covariance overlaps, computed using matrix_smooth().}

\item{t.traj}{Length of the MD simulation, in nanoseconds.}
}
\value{
A plot
}
\description{
This function generates a 2d plot of the diagonal of the smoothed covariance overlap matrix, generated by matrix_smooth().
The function generates a plot.
}
