% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpose_traj.R
\name{superpose_trj}
\alias{superpose_trj}
\title{Superpose an MD trajectory to its starting coordinates,
removing the roto-translational degrees of freedom}
\usage{
superpose_trj(str, traj, traj.format = "dcd")
}
\arguments{
\item{traj}{Trajectory file (dcd format)}

\item{topol}{Topology file (pdb format)}
}
\value{
superposed trajectory coordinates
}
\description{
Superpose an MD trajectory to its starting coordinates,
removing the roto-translational degrees of freedom
}
